provinces: u32 ids (no dropped-province overflow) + lakes split down the centre
Two fixes.

(1) u16 → u32 province ids. Province ids were u16 with 65535 as the NO_PROVINCE
sea sentinel, and the compaction assigned ids with `as u16`. A world that generated
>65534 provinces overflowed: ids wrapped and any province landing on 65535 rendered as
unowned "green" land — the empty regions seen on fine/large worlds. Widen `NO_PROVINCE`,
`Province.id`, `Province.neighbors`, `ProvinceBorder.neighbor`, the per-cell raster, the
RLE, the campaign `ProvinceDetail`/`ProvinceLive` ids, `prov_neighbors`, and the frontend
`ProvinceRaster` (Uint16Array → Uint32Array) to u32. Save-compatible: `migrate_raster_
sentinel` / `migrate_rle_sentinel` detect an old u16 blob (no u32::MAX present) on read and
remap its 65535 sea sentinel, so old worlds open correctly and the frontend always sees one
NO value. cargo check + 7 provinces:: tests + dynamics test + tsc all pass.

(2) Lakes split down the centre. Lakes were impassable holes (unowned). `assign_lakes_to_
nearest` runs last and fills each lake cell with its nearest-shore province via multi-source
BFS, so a wide lake shared by two provinces is divided along its medial axis — each side owns
up to the midline — instead of being a gap. Runs after all land-only stats, so it only
refines the raster/borders/hit-testing.

+++ src-tauri/src/sim/campaign/tick/disease.rs
@@ -211,7 +211,7 @@ impl CampaignSim {
                         if !neigh.is_empty() {
                             for h in 0..n {
                                 let hp = self.hub_province.get(h).copied().unwrap_or(-1);
-                                if hp >= 0 && neigh.contains(&(hp as u16)) {
+                                if hp >= 0 && neigh.contains(&(hp as u32)) {
                                     consider(self, h, &mut best);
                                 }
                             }
